Update: added Error boundary, React-router-dom and a 404 page for wild card catch
